Renamed some scripts to better associate the script function with the name.
New PCA analyzes over regions more segmented of the database, these new analyzes is more fast therefore they don't store the result in the database and show a plot in the  end of analisys.

It was make a attempt to increase the speed  to  create models   that prediect the  movie rates inside a genre, it was used Tensor Flow to make this.

It was created a new method of learning where each user has  a model to predict the movie rate that the user will give. The  average accuracy is between 87 to 94 percent for the new method, this very bigger than the old method ( script learning.py). The old method have a accuracy lower than majority classes therefore is not good.
It was created a new method of learning where each user has  a model to predict the movie rate that the user will give. The  average accuracy is between 87 to 94 percent for the new method, this very bigger than the old method ( script learning.py). The old method have a accuracy lower than majority classes therefore is not good.